Rename service module to follow projects convention
SDESK-7456
